Change struct to fit the libgit2 conventions
In particular, structs in libgit2 that get passed back through the API
to an application must be typdef'd in include/git2/types.h. This commit
will change the struct to be suited to this end; that is, rather than
typedef the struct as it's declared, we will typedef it in
include/git2/types.h.