git_diff() is working for all files in the repository.
Still have to check for new files on the filesystem. Currently tested it on
this repository, and it detected that diff.c was the only file to change.

Also, the is_file_changed() method may not be as efficant as it could be, but
it was easy to get running this way. If we have time we should try this by
comparing SHA1 hashes and see if that runs any faster.